turno-controller

+++ src/main/java/com/me/odontologo/service/implementation/TurnoService.java
@@ -2,6 +2,8 @@ package com.me.odontologo.service.implementation;
 
 import com.me.odontologo.entity.Turno;
 import com.me.odontologo.repository.ITurnoRepository;
+import com.me.odontologo.service.IOdontologoService;
+import com.me.odontologo.service.IPacienteService;
 import com.me.odontologo.service.ITurnoService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -11,19 +13,45 @@ import java.util.Optional;
 @Service
 public class TurnoService implements ITurnoService {
     private ITurnoRepository turnoRepository;
+    private IOdontologoService odontologoService;
+    private IPacienteService pacienteService;
 
     @Autowired
-    public TurnoService(ITurnoRepository turnoRepository) {
+    public TurnoService(ITurnoRepository turnoRepository,
+                        IOdontologoService odontologoService,
+                        IPacienteService pacienteService) {
         this.turnoRepository = turnoRepository;
+        this.odontologoService = odontologoService;
+        this.pacienteService = pacienteService;
     }
 
     @Override
     public Turno guardarTurno(Turno turno){
-        return turnoRepository.save(turno);
+        Turno turnoGuardado = null;
+        if (odontologoService.buscar(turno.getOdontologo().getId()).isPresent()
+                 && pacienteService.buscar(turno.getPaciente().getId()) != null){
+            turnoGuardado = turnoRepository.save(turno);
+        }
+        return turnoGuardado;
+    }
+
+    @Override
+    public Turno actualizarTurno(Turno turno){
+        Turno turnoActualizado = null;
+        if (turnoRepository.findById(turno.getId()).isPresent()){
+            turnoActualizado = turnoRepository.save(turno);
+        }
+        return turnoActualizado;
     }
     @Override
-    public void eliminarTurno(Long id){
-        turnoRepository.deleteById(id);
+    public boolean eliminarTurno(Long id){
+        boolean response = true;
+        if (turnoRepository.findById(id).isPresent()){
+            turnoRepository.deleteById(id);
+        }else{
+            response = false;
+        }
+        return response;
     }
     @Override
     public List<Turno> buscarTodosLosTurnos(){
